feat(be): 신고와 밴 추가

+++ backend/src/main/java/com/back/ourlog/domain/report/repository/ReportRepository.java
@@ -9,11 +9,11 @@ import org.springframework.data.repository.query.Param;
 
 import java.time.LocalDateTime;
 
-public interface ReportRepository extends JpaRepository<Report, Long> {
+public interface ReportRepository extends JpaRepository<Report, Integer> {
 
     boolean existsByReporterAndTargetAndType(User reporter, User target, ReportReason type);
 
     @Query("SELECT COUNT(r) FROM Report r WHERE r.target.id = :userId AND r.reportedAt >= :since")
-    long countRecentReportsForUser(@Param("userId") Long userId, @Param("since") LocalDateTime since);
+    long countRecentReportsForUser(@Param("userId") Integer userId, @Param("since") LocalDateTime since);
 }
